+++ hemi/internal/web_proxy_http1.go
@@ -90,7 +90,7 @@ func (h *http1Proxy) Handle(req Request, resp Response) (next bool) { // forward
 	// TODO: use stream1.ForwardProxy() or stream1.ReverseProxy()
 
 	req1 := stream1.Request()
-	if !req1.copyHeadFrom(req, h.hostname, h.colonPort, h.viaName) {
+	if !req1.copyHeadFrom(req, h.hostname, h.colonPort, h.viaName, h.addRequestHeaders, h.delRequestHeaders) {
 		stream1.markBroken()
 		resp.SendBadGateway(nil)
 		return